# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8070","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getCountries","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"code","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}},"post":{"tags":["country-controller"],"operationId":"saveCountries","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveCountriesRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{language}/language":{"get":{"tags":["country-controller"],"operationId":"getCountriesByLanguage","parameters":[{"name":"language","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{continent}/continent":{"get":{"tags":["country-controller"],"operationId":"getCountriesByContinent","parameters":[{"name":"continent","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"getCountriesWithMostBorders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}}},"components":{"schemas":{"SaveCountriesRequestDTO":{"type":"object","properties":{"amountOfCountryToSave":{"type":"integer","format":"int64"}}},"CountryDTO":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}}}}}